requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
boto3==1.9.159
botocore==1.12.159
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
docutils==0.14
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.8
iso3166==1.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.0.0
numpy==1.16.3
pandas==0.24.2
pluggy==0.11.0
prettytable==0.7.2
py==1.8.0
pyaml==18.11.0
pyparsing==2.4.0
pyrsistent==0.15.2
pytest==4.0.1
python-dateutil==2.8.0
python-whois==0.7.1
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-file==1.4.3
s3transfer==0.2.0
scikit-learn==0.21.1
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.1
tldextract==2.2.1
urllib3==1.25.2
validator-collection==1.3.5
Werkzeug==0.15.4